# Use performance optimized shuffle_do() method