--- a/examples/bank_reserves/bank_reserves/model.py
+++ b/examples/bank_reserves/bank_reserves/model.py
@@ -152,7 +152,7 @@ class BankReserves(mesa.Model):
 
     def step(self):
         # tell all the agents in the model to run their step function
-        self.agents.shuffle().do("step")
+        self.agents.shuffle_do("step")
         # collect data
         self.datacollector.collect(self)