--- a/examples/charts/charts/model.py
+++ b/examples/charts/charts/model.py
@@ -135,7 +135,7 @@ class Charts(mesa.Model):
 
     def step(self):
         # tell all the agents in the model to run their step function
-        self.agents.shuffle().do("step")
+        self.agents.shuffle_do("step")
         # collect data
         self.datacollector.collect(self)