--- a/examples/el_farol/el_farol/model.py
+++ b/examples/el_farol/el_farol/model.py
@@ -34,9 +34,9 @@ class ElFarolBar(mesa.Model):
     def step(self):
         self.datacollector.collect(self)
         self.attendance = 0
-        self.agents.shuffle().do("update_attendance")
+        self.agents.shuffle_do("update_attendance")
         # We ensure that the length of history is constant
         # after each step.
         self.history.pop(0)
         self.history.append(self.attendance)
-        self.agents.shuffle().do("update_strategies")
+        self.agents.shuffle_do("update_strategies")