--- a/examples/forest_fire/Forest Fire Model.ipynb
+++ b/examples/forest_fire/Forest Fire Model.ipynb
@@ -149,7 +149,7 @@
     "        \"\"\"\n",
     "        Advance the model by one step.\n",
     "        \"\"\"\n",
-    "        self.agents.shuffle().do(\"step\")\n",
+    "        self.agents.shuffle_do(\"step\")\n",
     "        self.dc.collect(self)\n",
     "        # Halt if no more fire\n",
     "        if self.count_type(self, \"On Fire\") == 0:\n",