--- a/examples/virus_on_network/virus_on_network/model.py
+++ b/examples/virus_on_network/virus_on_network/model.py
@@ -95,7 +95,7 @@ class VirusOnNetwork(mesa.Model):
             return math.inf
 
     def step(self):
-        self.agents.shuffle().do("step")
+        self.agents.shuffle_do("step")
         # collect data
         self.datacollector.collect(self)