Replace `shuffle().do()` in 18 models with the performance optimized `shuffle_do()` method.

`shuffle_do()` was added in https://github.com/projectmesa/mesa/pull/2283 and is first included in the upcoming Mesa `3.0.0a5` release. It allows for a drop-in performance upgrade:

```diff
- self.agents.shuffle().do("step")
+ self.agents.shuffle_do("step")
```
Which can speed up the Agent activation by over 3x depending on the scenario.

| Configuration                | shuffle().do() | shuffle(inplace=True).do() | shuffle_do() |
|------------------------------|----------------|----------------------------|--------------|
| 10,000 agents, 1,000 steps   | 8.27 s         | 3.65 s                     | 3.06 s       |
| 100,000 agents, 100 steps    | 13.71 s        | 6.31 s                     | 3.71 s       |
| 1,000,000 agents, 10 steps   | 18.36 s        | 9.44 s                     | 5.75 s       |

`shuffle_do()` can be used to replace the old `RandomActivation` scheduler, as demonstrated in https://github.com/projectmesa/mesa-examples/pull/183.

Part of https://github.com/projectmesa/mesa-examples/issues/111.